+++ src/@demo/demo2/index.js
@@ -3,18 +3,18 @@ import './styles.css';
 export default (ColorTranslator) => {
 
     const container = document.createElement('div');
-    const hsl = { h: 0, s: '100%', l: '50%' };
+    const color = new ColorTranslator({ r: 255, g: 0, b: 0 });
     const hue = [0, 30, 60, 120, 240, 280, 320, 0];
     let rainbow;
 
     for (let r = 0; r < hue.length; r++) {
 
-        hsl.h = hue[r];
+        color.setH(hue[r]);
 
-        const rgb = ColorTranslator.toHEX(hsl);
         const bow = document.createElement('div');
+
         bow.classList.add('rainbow');
-        bow.style.background = rgb;
+        bow.style.background = color.HEX;
 
         if (rainbow) {
             rainbow.appendChild(bow);