+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.1.0] - 2019-07-12
+
+### Added
+
+- Add the possibility to instantiate the class using any of the inputs
+- Add 11 public chainable methods and 25 public readonly properties
+- Update some demo examples to use the new class, methods, and properties
+- Solve a bug with webpack and the clean-webpack-plugin package
+- Fix a bug with RGBA and HSLA CSS output
+- Change the CSS CMYK output from `device-cmyk` to `cmyk`
+
 ## [1.0.5] - 2019-07-05
 
 ### Added